backend/requirements.txt
# --- Framework ---
fastapi
uvicorn[standard]

# --- Database ---
sqlalchemy
psycopg2-binary
alembic

# --- Authentication ---
python-jose[cryptography]
passlib[bcrypt]

# --- File Storage ---
supabase

# --- Resume Parsing & NLP ---
pymupdf
spacy

# --- Embeddings & Vector Search ---
sentence-transformers
faiss-cpu
tf-keras

# --- AI Orchestration & Google Integration ---
langchain
langchain-community
langchain-google-genai

# --- Utilities ---
python-dotenv
numpy
python-multipart
